Merge branch 'master' into brewing-recipe

+++ src/main/java/net/mcreator/element/types/Block.java
@@ -100,7 +100,7 @@ import java.util.Map;
 
 	public boolean isNotColidable;
 	public StepSound soundOnStep;
-	public double luminance;
+	public int luminance;
 	public boolean unbreakable;
 	public int breakHarvestLevel;
